backend/requirements.txt
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.0.0
google-genai>=0.2.0
pydantic>=2.0.0
pydantic-settings>=2.0.0
python-dotenv>=1.0.0
speechrecognition>=3.10.0
pydub>=0.25.1
edge-tts>=6.1.0

# New dependencies for scoring system
sqlalchemy>=2.0.0
reportlab>=4.0.0
